requirements.txt
opencv-python 
mediapipe==0.10.8
Pillow==10.0.1
numpy 
torch 
torchvision 
torchsummary


